Moved RegisterActivity to signIn
