+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SwiftWallet">
         <activity
-            android:name="com.sdp.swiftwallet.presentation.RegisterActivity"
+            android:name="com.sdp.swiftwallet.presentation.signIn.RegisterActivity"
             android:exported="false" />
         <activity
             android:name="com.sdp.swiftwallet.TransactionHistoryActivity"
